services/api_gateway/requirements.txt
asgiref==3.10.0
Django==4.2.25
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
redis==6.4.0
sqlparse==0.5.3
